Updated XPath
eFukt changed up their website and the XPath needed to be updated in my
channel. Also got rid of using RegEx in the ServiceCode.pys file, and
only use XPath. Also, will be pushing a release shortly, and dropping
the release versioning I've been doing in the commits (didn't realize
that releases existed on GitHub).